refactor ShadowFontConfiguration
