+++ carddrawer/src/main/java/com/meli/android/carddrawer/configuration/shadow/ShadowFontConfiguration.java
@@ -1,4 +1,4 @@
-package com.meli.android.carddrawer.configuration;
+package com.meli.android.carddrawer.configuration.shadow;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
@@ -6,18 +6,19 @@ import android.support.v4.content.ContextCompat;
 import android.text.TextPaint;
 import com.meli.android.carddrawer.R;
 
-class ShadowConfiguration {
+public class ShadowFontConfiguration implements ShadowConfiguration {
     private final int shadowRadius;
     private final int shadowColor;
     private final float shadowHeight;
 
-    ShadowConfiguration(@NonNull final Context context) {
+    public ShadowFontConfiguration(@NonNull final Context context) {
         shadowRadius = context.getResources().getDimensionPixelSize(R.dimen.card_drawer_shadow_radius);
         shadowColor = ContextCompat.getColor(context, R.color.card_drawer_number_shadow_color);
         shadowHeight = (context.getResources().getDisplayMetrics().density * 1) + 0.5f;
     }
 
-    void drawShadow(@NonNull final TextPaint textPaint) {
+    @Override
+    public void drawShadow(@NonNull final TextPaint textPaint) {
         textPaint.setShadowLayer(shadowRadius, 0, shadowHeight, shadowColor);
     }
 }
