+++ carddrawer/src/test/java/com/meli/android/carddrawer/configuration/LightFontConfigurationTest.java
@@ -7,6 +7,7 @@ import android.text.TextPaint;
 import com.meli.android.carddrawer.BasicRobolectricTest;
 import com.meli.android.carddrawer.R;
 
+import com.meli.android.carddrawer.configuration.shadow.ShadowFontConfiguration;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -27,8 +28,8 @@ public class LightFontConfigurationTest extends BasicRobolectricTest {
 
     @Before
     public void doBefore() {
-        ShadowConfiguration shadowConfiguration = new ShadowConfiguration(getContext());
-        lightFontConfiguration = new LightFontConfiguration(getContext(), shadowConfiguration);
+        ShadowFontConfiguration shadowFontConfiguration = new ShadowFontConfiguration(getContext());
+        lightFontConfiguration = new LightFontConfiguration(getContext(), shadowFontConfiguration);
     }
 
     @Test
